advent day 2: haskell
This one took longer than I wanted it to, mostly due to the
flexibility of the language and operator precedence. About half of
the way through I considered dropping it, as picking up Haskell in
a day seemed ridiculous. By the end of it, something finally
clicked where I was actually a bit disappointed that I wouldn't be
able to use this on the next day's challenges. The problem itself
was perfect for the language and the way it wants you to think.

My biggest complaint of Haskell is that in order to do IO, you get
thrust into monad land, which increases the complexity of learning
significantly. Suddenly having to deal with operator precedence
with >>= and $ and . was a bit much. I could have pasted the
problem set into the program, but for some reason I'm all about
insane ramp-up difficulty. Would definitely play with the language
again.