advent day 3: nim
This was an interesting one! I never really had much interest in
trying to pick up nim, but I was overall pleasantly surprised. I'm
not sure I'd reach for it in a lot of cases, but it's pretty clear
that it came from Python inspiration as advertised.

Putting together a quick program was trivial, which is always nice,
and it was nice to know that the quick program you put together isn't
totally non-idiomatic, which I find is often the case when you put
something simple together vs. a legitimate piece of software.

Favorite thing about nim: first class iterators! I love iterators
to the moon and back, so being able to build a yielding procedure
with some first class semantics was coolio. That set me up for some
joy, and despite my (much) longer list of complaints, I still enjoyed
them a lot.

Another nice thing: the syntax. By and large the nim syntax hits a
sweet spot between being both easy to read _and_ write. Again, that
Python influence is there. There are some cases where nim decided to
keep the keyword footprint low by using weird curly-dot suffixes to
signatures to indicate things like pure functions, exceptions that
are raised, etc. That part is... odd. I don't think I hate it though.

Least favorite: nim is quirky for a new language. It makes a lot of
weird decisions that I'd expect from a language that had grown
organically over decades: not something I'd want in a new language.

- There are too many ways to return. Explicit, implicit last statement,
  and "magic" variable assignment.
- Exception handling is a pattern I don't like in general -- I much
  prefer the way Rust handles it. Or even Go, which usually lets you
  just tack on extra return values. Nim doesn't really have constructs
  to handle that kind of pattern elegantly. Even more odd, nim seems
  to arbitrarily decide to use boolean success return values or
  exceptions somewhat arbitrarily. (See file opening.)
  All that being said, I _do_ like that you're forced to handle
  exceptions, using explicit throwaway keywords like `discard`. I like
  that more than say, Java, where there's no explicit way to do so, and
  IDEs just yolo if you name the variable a certain way.
- Iterators move in weird ways, they'll only mark themselves as ended
  after you make one extra call _past_ the end, which means you get a
  garbage value you have to know to drop. You also can't ever simply
  iterate over the values... you often do a "while true" loop to wrap
  the bad value and prevent your procedure from returning the garbage.
  This is documented in the nim docs and is absolutely bonkers.

All in all, I think nim would be a very interesting replacement for
what you may typically think to write as a shell script. It's nice and
scripty and reads nicely. There's way too much quirkiness in it that
would probably prevent me from recommending it for production use
outside that.