advent day 4: julia
Indexing at 1, really? Otherwise, oh my goodness, what a pleasure
Julia was to work with! At first I thought this may be a stretch to
accomplish, but it turns out this is pretty much the job it is
designed for. Taking a bunch of input, splitting it up, vectorizing
the work to be done... it was all super trivial and a nice alternative
to functional mapping patterns (though those are there too, where it
makes sense).

I grabbed a bunch of different tools off the shelf in Julia for the
various validation pieces just to try them out (is it cheating if you
just reach for a regex library in every language you pick up?). 10/10,
would definitely consider for any data related needs I may have.