advent day 5: shakespeare
So I had a few other languages in mind to try out, but when I saw
that this problem had very little in the means of being able to
leverage them... I decided to take a break and drink some whiskey.

One dram of whiskey later, and I decided to punch out to a novelty
language to change it up. I pretty quickly settled on Shakespeare
due to just how ridiculous it was, and how nice the results actually
were (as far as the goal of looking like actual Shakespeare plays).

Essentially, this is just a language that mimics assembly and gives
you basic ability to manipuate powers of two, do some simple math
operations... and that's about it. You get about by bringing players
in and out, which serve as registers. Communicating with them allows
you to push/pull data and do simple conditionals.

Due to the limits of the interpreter I was using, I couldn't
actually use the language to do the _whole_ problem, but the rest of
the problem was a simple sort, so I didn't feel to bad busting out
Julia from yesterday to play with some more. Shakespeare was so slow
I even got a chance to use the parallel directives in Julia to thread
up the workload. Nice!